RDKBDEV-1838: fix incorrect broadcast address in active lan
Reason for change:
fix incorrect broadcast address in active lan interfaces ( Part of OFW-3610 )

When brlan0 comes up after setting Device.IP.Interface.4.Enable to true, It has
broadcast address of 225.168.1.1 for 192.168.1.1/24 because the bit order got
reversed when doing bitwise OR.

After setting the dml to true, ensured the broadcast as 192.168.1.255 for the
address 192.168.1.1/24

Test Procedure: Sanity.
Risks: None.

Signed-off-by: akashdhanabal <adhanabal.ext@libertyglobal.com>
Change-Id: I5ef324bde076ea0b25a05683d559cc57efbec6db